+++ src/call/mod.rs
@@ -750,7 +750,7 @@ impl SinkBase {
         // temporary fix: buffer hint with send meta will not send out any metadata.
         // note: only the first message can enter this code block.
         if self.send_metadata {
-            ser(t, &mut self.buffer);
+            ser(t, &mut self.buffer)?;
             self.buf_flags = Some(flags);
             self.start_send_buffer_message(false, call, call_flags)?;
             self.send_metadata = false;
@@ -763,7 +763,7 @@ impl SinkBase {
             self.start_send_buffer_message(true, call, call_flags)?;
         }
 
-        ser(t, &mut self.buffer);
+        ser(t, &mut self.buffer)?;
         let hint = flags.get_buffer_hint();
         self.last_buf_hint &= hint;
         self.buf_flags = Some(flags);